probable fix for hanging while cloning an illuminant profile due to nc*w_l having a higher value than 255
svn path=/trunk/koffice/; revision=1098746

+++ krita/plugins/extensions/painterlyframework/kis_illuminant_profile.cpp
@@ -174,7 +174,7 @@ KisIlluminantProfile::KisIlluminantProfile(const KisIlluminantProfile &copy)
 
         nc = copy.nc;
         coeffs = new double[nc*m_wl];
-        for (qint8 i = 0; i < nc*m_wl; i++)
+        for (int i = 0; i < nc*m_wl; i++)
             coeffs[i] = copy.coeffs[i];
 
         m_refvec = new double[m_wl];
